fix(ci): Repair CI for OS contributors (#91)
This PR fixes the CI workflow to generate values for environment variables if the MOCK env var is set. Previously the values needed to be set on the repository which works as well, but exists outside the code and elevated permissions to just run CI (meaning OS contributions would fail CI).

This behaviour is fine since it doesn't actually use these variables as network requests are stubbed when MOCK is set.

This was achieved by using the current repo for generating an image, rather than the published version. If the local image passes all tests and has been reviewed, it makes sense for it to be published afterward. Testing changes in this repo against the published version prevents testing new behaviour.