feat: Add link to new release issue creation
In order to make sure that we follow all steps to create a release I've created a new template and this makes reference to it.

I've merged the template to `master` without thinking about it. Please have a look at it and let me know if you want some tweaks:
https://github.com/getsentry/action-release/commit/8b4720d3d1512cfb105c59518acd3bd52c6323e6