feat: implement 'working-directory' argument
When running the action-release Github Action, occasionally one
might need to perform auto-detection of associated commits in a
directory other than the job's working directory, such as in the
case where an action performs multiple checkouts. As Github
Actions do not, by default, support the ability to run an action
in any directory other than the job's working directory, this
change implements an argument for specifying which directory to
gather sentry release information from.