chore(release): Add craft publish workflow
Previously, any push to `master` built and uploaded the action's underlying
docker image to the GitHub registry under the `latest` tag. That meant,
regardless of version specified in a user's workflow they'd get the latest
change.

The introduction of craft allows us to push versioned images to the registry and
associate a release's action with that version.

At the moment, this workflow does not incorporate tags for major (e.g. `v1`) and
major.minor (e.g. `v1.2`). The ability to publish several git tags will probably
have to be added to craft or publish directly.

Closes: #129